requirements.txt
fastapi==0.115.12
psycopg2-binary==2.9.9
python-dotenv==1.0.1
sqlalchemy==2.0.38
uvicorn==0.34.0
pydantic>=2.0.2,<3.0.0
pydantic[email]>=2.0.2,<3.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.3.0
cryptography==42.0.0
python-dateutil==2.8.2
pytz==2023.3
